requirements.txt
#python3.13
requests
jinja2
WTForms
Pillow
PyMySQL
